Query Functions to be reviewed
Update and Delete function for SQLite query has been updated and the same has been developed with unit test. The Update functionality provided as 1st point in backlogs are being tried, so those vi's are also uploaded in this.
All other functionalities are already Checked which normally Run before starting to develop the above said things for review.